[AOTI] Improve the two-pass wrapper codegen
Summary: For the second-pass, we don't have to rerun the whole inductor flow again. This PR moves that second-pass to the codegen time. This change not only speeds up the compilation, but also removes kernel scheduling inconsistency between the two passes. Another future improvement is to make the second-pass reuse the scheduler and do the wrapper codegen only.

ghstack-source-id: 61eb5f0527b75ba5c60fc0af9814f5bac83ca265
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113762